requirements.txt
accelerate==0.29.3
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
appscript==1.2.5
argcomplete==3.3.0
asttokens==2.4.1
attrs==23.2.0
azure-core==1.30.1
azure-monitor-opentelemetry-exporter==1.0.0b25
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptography==42.0.5
cymem==2.0.8
dataclasses-json==0.6.6
decorator==4.4.2
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docutils==0.21.2
e2b==0.15.0
e2b-code-interpreter==0.0.3
executing==2.0.1
fastapi==0.110.2
filelock==3.13.4
filetype==1.2.0
fixedint==0.1.6
Flask==3.0.3
Flask-Cors==4.0.0
flask-restx==1.3.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google_search_results==2.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
groq==0.5.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipython==8.24.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpath-ng==1.6.1
jsonpointer==2.4
jsonrpcclient==4.0.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.1
langchain-core==0.2.1
langchain-openai==0.1.7
langcodes==3.4.0
langgraph==0.0.54
langsmith==0.1.60
language_data==1.2.0
lazy_loader==0.4
llama-index==0.10.37
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.37.post1
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.20
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.22
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.3
llamaindex-py-client==0.1.19
lxml==5.2.1
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
more-itertools==10.2.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.30.1
opencv-python==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
ply==3.11
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.43
promptflow==1.10.0
promptflow-core==1.10.0
promptflow-devkit==1.10.0
promptflow-tools==1.4.0
promptflow-tracing==1.10.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydash==7.0.7
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc-core==10.2
pyobjc-framework-ApplicationServices==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-CoreText==10.2
pyobjc-framework-Quartz==10.2
pyparsing==3.1.2
pypdf==4.2.0
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
rubicon-objc==0.4.8
safetensors==0.4.3
schedule==1.2.1
scikit-image==0.23.1
scipy==1.13.0
selenium==4.19.0
setuptools==69.2.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
strictyaml==1.7.3
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tavily-python==0.3.3
tenacity==8.3.0
thinc==8.2.3
tifffile==2024.2.12
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uuid6==2024.1.12
uvicorn==0.29.0
waitress==2.1.2
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
